ainsights-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.1
scikit-learn==1.5.2
joblib==1.4.2
